cài đặt menu repo

+++ src/main/java/faculty/ntu/cms/repositories/MenuItemRepository.java
@@ -4,6 +4,7 @@ import faculty.ntu.cms.models.MenuItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface MenuItemRepository extends JpaRepository<MenuItem, Integer> {
     // Lấy danh sách menu theo tên, đang hiển thị, có sắp xếp
@@ -11,5 +12,6 @@ public interface MenuItemRepository extends JpaRepository<MenuItem, Integer> {
     List<MenuItem> findByMenuNameAndActiveTrueOrderByOrder(String menuName);
     // Lấy danh sách menu con theo menu cha
     // Dùng để hiển thị menu cấp 2, cấp 3
-    List<MenuItem> findByParentId(Long parentId);
+    List<MenuItem> findByParentId(int parentId);
+
 }
